fix: address Cosmos PR review findings
Split the driver under the line limit and harden feed pagination, filters,
credentials, HTTPS, throttling retries, and filtered totals.
